[#36] Move node attribute code into separate file
